Basic Check List Complete
This is the base program being complete. The next part is to implement script checking and creating a generic script as shown in the flowchart purple block. I also need to implement a good reliable way to check for root access. I guess checking if the SU binary exists is a more reliable way to check for root access rather than checking the props. Eh good night, I'll work on this later.