QOL Changes
I never thought about multiple reports. My bad. Now it creates a new folder for each report. I will be changing a lot of the code to store the data in a json file instead of variables (which I should have done from the starting but it is what it is)